feat(engine): add MSET/MGET batch commands
Add MSET and MGET commands for writing and reading multiple blocks in a
single round-trip. Eliminates N×RTT overhead for bulk operations over
high-latency connections.

Also includes:
- constant-time AUTH token comparison
- internal error messages no longer leak filesystem paths to clients
- thread-safe CRC32 table initialization
- fix signed integer overflow (UB) in ReadLe32 for high bytes
- async-signal-safe signal handler via watcher thread
- correct IPv6 bracketed address parsing in URI parser
- CRLF stripping protection in Go client command serialization

+++ src/chunk_store.cpp
@@ -737,11 +737,10 @@ std::uint16_t ReadLe16(const std::vector<std::uint8_t>& data, std::size_t offset
 }
 
 std::uint32_t ReadLe32(const std::vector<std::uint8_t>& data, std::size_t offset) {
-    return static_cast<std::uint32_t>(
-        static_cast<std::uint32_t>(data[offset]) |
-        static_cast<std::uint32_t>(data[offset + 1] << 8U) |
-        static_cast<std::uint32_t>(data[offset + 2] << 16U) |
-        static_cast<std::uint32_t>(data[offset + 3] << 24U));
+    return static_cast<std::uint32_t>(data[offset]) |
+           (static_cast<std::uint32_t>(data[offset + 1]) << 8U) |
+           (static_cast<std::uint32_t>(data[offset + 2]) << 16U) |
+           (static_cast<std::uint32_t>(data[offset + 3]) << 24U);
 }
 
 std::uint64_t ReadLe64(const std::vector<std::uint8_t>& data, std::size_t offset) {
